requirements.txt
Django>=4.1,<5.0
djangorestframework>=3.13,<4.0
gunicorn>=20.1,<21.0
psycopg2>=2.9,<3.0
pillow>=9.0,<10.0
torch>=1.13,<2.0
torchvision>=0.14,<1.0
